+++ zio-flow/shared/src/main/scala/zio/flow/Mappable.scala
@@ -1,6 +1,6 @@
 package zio.flow
 
-import zio.flow.remote.Remote
+import zio.flow.Remote
 
 sealed trait Mappable[F[_]] {
   def performMap[A, B](fa: Remote[F[A]], ab: Remote[A] => Remote[B]): Remote[F[B]]
