+++ zio-flow/shared/src/main/scala/zio/flow/Tuples.scala
@@ -1,7 +1,5 @@
 package zio.flow
 
-import zio.flow.remote.Remote
-
 sealed trait Tuple[Z] {
   type out
