+++ zio-flow/shared/src/main/scala/zio/flow/remote/RemoteExecutingFlowSyntax.scala
@@ -1,7 +1,6 @@
 package zio.flow.remote
 
-import zio.flow.zFlow.ZFlow
-import zio.flow.{ ActivityError, ExecutingFlow, FlowId }
+import zio.flow.{ ActivityError, ExecutingFlow, FlowId, Remote, ZFlow }
 
 class RemoteExecutingFlowSyntax[A](self: Remote[A]) {
