+++ zio-flow/shared/src/main/scala/zio/flow/remote/RemoteFractionalSyntax.scala
@@ -1,5 +1,6 @@
 package zio.flow.remote
 
+import zio.flow._
 import zio.schema.Schema
 
 class RemoteFractionalSyntax[A](self: Remote[A]) {
