+++ zio-flow/shared/src/main/scala/zio/flow/remote/RemoteMappableSyntax.scala
@@ -1,6 +1,6 @@
 package zio.flow.remote
 
-import zio.flow.Mappable
+import zio.flow.{ Mappable, Remote }
 
 class RemoteMappableSyntax[A](self: Remote[A]) {
