+++ zio-flow/shared/src/main/scala/zio/flow/remote/RemoteNumericSyntax.scala
@@ -1,5 +1,7 @@
 package zio.flow.remote
 
+import zio.flow._
+
 class RemoteNumericSyntax[A](self: Remote[A]) {
 
   final def +(that: Remote[A])(implicit numeric: Numeric[A]): Remote[A] =
