+++ zio-flow/shared/src/main/scala/zio/flow/remote/RemoteStringSyntax.scala
@@ -1,5 +1,7 @@
 package zio.flow.remote
 
+import zio.flow._
+
 class RemoteStringSyntax(self: Remote[String]) {
   def length: Remote[Int] = Remote.Length(self)
 }
