+++ zio-flow/shared/src/main/scala/zio/flow/remote/numerics.scala
@@ -1,5 +1,6 @@
 package zio.flow.remote
 
+import zio.flow.Remote
 import zio.schema.Schema
 
 sealed trait Numeric[A] {
