+++ zio-flow/shared/src/main/scala/zio/flow/server/PersistentExecutor.scala
@@ -5,10 +5,9 @@ import java.io.IOException
 import zio._
 import zio.clock._
 import zio.flow._
-import zio.flow.remote.{ Remote, SchemaAndValue }
 import zio.flow.server.PersistentExecutor.{ State, TState }
-import zio.flow.zFlow.ZFlow._
-import zio.flow.zFlow.{ ZFlow, ZFlowExecutor }
+import zio.flow.ZFlow._
+import zio.flow.zFlow.ZFlowExecutor
 import zio.schema.Schema
 
 final case class PersistentExecutor(
