+++ zio-flow/shared/src/main/scala/zio/flow/zFlow/ZFlowExecutor.scala
@@ -5,8 +5,8 @@ import java.time.Duration
 import zio._
 import zio.clock.Clock
 import zio.console.putStrLn
-import zio.flow.remote.Remote
-import zio.flow.{ ActivityError, ExecutingFlow, OperationExecutor }
+import zio.flow.Remote
+import zio.flow.{ ActivityError, ExecutingFlow, OperationExecutor, ZFlow }
 import zio.schema.Schema
 
 trait ZFlowExecutor[-U] {
