+++ zio-flow/shared/src/test/scala/zio/flow/GoodcoverUseCase.scala
@@ -4,9 +4,7 @@ import java.net.URI
 
 import zio.flow
 import zio.flow.ZFlowMethodSpec.setBoolVarAfterSleep
-import zio.flow.remote.{ Remote, RemoteVariable }
 import zio.flow.utils.ZFlowAssertionSyntax.InMemoryZFlowAssertion
-import zio.flow.zFlow.ZFlow
 import zio.schema.{ DeriveSchema, Schema }
 import zio.test.Assertion.equalTo
 import zio.test._
