+++ zio-flow/shared/src/test/scala/zio/flow/RemoteInstantSpec.scala
@@ -1,6 +1,5 @@
 package zio.flow
 
-import zio.flow.remote.{ Remote, _ }
 import zio.flow.utils.RemoteAssertionSyntax.RemoteAssertionOps
 import zio.test._
