+++ zio-flow/shared/src/test/scala/zio/flow/ZFlowExecutorSpec.scala
@@ -3,10 +3,8 @@ package zio.flow
 import java.net.URI
 import java.time.Instant
 
-import zio.flow.remote.{ Remote, _ }
 import zio.flow.utils.ZFlowAssertionSyntax.InMemoryZFlowAssertion
 import zio.flow.utils.ZFlowAssertionSyntax.Mocks.mockInMemoryTestClock
-import zio.flow.zFlow.ZFlow
 import zio.flow.zFlow.ZFlowExecutor.InMemory.{ CompileStatus, State, TState }
 import zio.schema.DeriveSchema.gen
 import zio.schema.Schema
