+++ zio-flow/shared/src/test/scala/zio/flow/ZFlowMethodSpec.scala
@@ -3,10 +3,8 @@ package zio.flow
 import zio.clock.Clock
 import zio.console.Console
 import zio.duration.durationInt
-import zio.flow.remote._
 import zio.flow.utils.ZFlowAssertionSyntax.InMemoryZFlowAssertion
 import zio.flow.utils.ZFlowAssertionSyntax.Mocks.mockActivity
-import zio.flow.zFlow.ZFlow
 import zio.schema.Schema
 import zio.test.Assertion.equalTo
 import zio.test._
