+++ zio-flow/shared/src/test/scala/zio/flow/utils/RemoteAssertionSyntax.scala
@@ -1,6 +1,6 @@
 package zio.flow.utils
 
-import zio.flow.remote.Remote
+import zio.flow._
 import zio.test.Assertion.{ equalTo, isRight }
 import zio.test.{ TestResult, assert }
